% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_VB_DirBer.R
\name{algo_VB_DirBer_R}
\alias{algo_VB_DirBer_R}
\title{Collapsed VB for the Dirichlet-Bernoulli model}
\usage{
algo_VB_DirBer_R(V, Z, K, alpha = 1, beta = 1, gamma = 1, iter = 100)
}
\description{
Collapsed Variational Inference under the Truncated Symmetric Dirichlet Prior
Instead of computing the real limit K -> inf, we keep the origial Dirichlet-Multinomial
with a large K and Dirichlet parameters alpha/K
}
