% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BernoulliNMF.R
\name{BernoulliNMF}
\alias{BernoulliNMF}
\title{Collapsed Gibbs Sampler for Bernoulli models}
\usage{
BernoulliNMF(V, K = K, model = "DirBer", method = "Gibbs", alpha = 1,
  beta = 1, gamma = 1, iter = 1000, burnin = 0.5)
}
\arguments{
\item{V}{Observed matrix with FxN dimensions}

\item{K}{Number of latent dimensions}

\item{alpha}{parameter for the Beta distribution}

\item{beta}{parameter for the Beta distribution}

\item{gamma}{parameter for the Dirichlet distribution}

\item{iter}{number of Gibbs samples}

\item{burnin}{proportions of initial samples that are ignored for burnin}
}
\value{
FxNxJ arrays with samples from the variables
}
\description{
Samples the auxiliary variables (Z or Z,C) using a collapsed
Gibbs sampler.
}
