% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_Gibbs_DirBer.R
\name{expectation.GibbsDirBer}
\alias{expectation.GibbsDirBer}
\title{Expectation in Dirichlet Bernoulli model}
\usage{
\method{expectation}{GibbsDirBer}(modelGibbsDirBer)
}
\description{
Compute expected matrix V
}
\details{
It gets V=E[WH] given each sample from Z
}
