% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commons.R
\name{expectation.VBBernoulli}
\alias{expectation.VBBernoulli}
\title{Expectation of Variational Inference model}
\usage{
\method{expectation}{VBBernoulli}(modelVB)
}
\arguments{
\item{E_W}{F x K matrix with expected dictionary W}

\item{E_H}{N x K matrix with expected activations H}
}
