% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_Gibbs_DirBer.R
\name{loglikelihood_old.GibbsDirBer}
\alias{loglikelihood_old.GibbsDirBer}
\title{Dirichlet-Bernoulli loglikelihood}
\usage{
loglikelihood_old.GibbsDirBer(modelGibbsDirBer, V)
}
\description{
Compute likelihood of some matrix V  p(V | V)
}
\details{
It uses the samples from Z and p(V | V) = E[p(V | Z)]
}
