% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCEM - DirBer.R
\name{mmle_dirber_mcem}
\alias{mmle_dirber_mcem}
\title{NMF with MMLE}
\usage{
mmle_dirber_mcem(V, initW, alpha = 1, maxiters = 100, nsamples = 50,
  burnin = 0.5)
}
\arguments{
\item{V}{matrix to be factorized}

\item{maxiters}{maximum number of iterations}

\item{nsamples}{number of samples per iteration in the E-step}

\item{K}{number of latent factors (or dimensions)}

\item{W}{initial matrix W}

\item{mc}{mcem or saem.}
}
\value{
W last estimator of W
        W_traces a K x samples - matrix vector with the evolution of row norms
}
\details{
The number of iterations is set to maxiters.
In the future, there should be a convergence check in case
the KL converges before maxiters.
}
