% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_VB_DirBer.R
\name{VB_DirBer}
\alias{VB_DirBer}
\title{Collapsed VB for the Dirichlet-Bernoulli model}
\usage{
VB_DirBer(V, Z, K, alpha = 1, beta = 1, gamma = 1, iter = 100)
}
\description{
Collapsed Variational Inference under the Truncated Symmetric Dirichlet Prior
Instead of computing the real limit K -> inf, we keep the origial Dirichlet-Multinomial
with a large K and Dirichlet parameters alpha/K
}
